bump go, bump deps for cve, bump opencost and move to opencost/core Signed-off-by: Cliff Colvin <ccolvin@kubecost.com>
