+++ pkg/cmd/aggregatedcommandbuilder.go
@@ -12,7 +12,7 @@ import (
 	"github.com/kubecost/kubectl-cost/pkg/cmd/display"
 	"github.com/kubecost/kubectl-cost/pkg/cmd/utilities"
 	"github.com/kubecost/kubectl-cost/pkg/query"
-	"github.com/opencost/opencost/pkg/log"
+	"github.com/opencost/opencost/core/pkg/log"
 )
 
 type AggregatedAllocationOptions struct {
