+++ pkg/cmd/tui.go
@@ -14,8 +14,8 @@ import (
 	"github.com/kubecost/kubectl-cost/pkg/cmd/display"
 	"github.com/kubecost/kubectl-cost/pkg/cmd/utilities"
 	"github.com/kubecost/kubectl-cost/pkg/query"
-	"github.com/opencost/opencost/pkg/kubecost"
-	"github.com/opencost/opencost/pkg/log"
+	"github.com/opencost/opencost/core/pkg/log"
+	"github.com/opencost/opencost/core/pkg/opencost"
 	"github.com/rivo/tview"
 	"github.com/spf13/cobra"
 )
@@ -213,7 +213,7 @@ func runTUI(ko *utilities.KubeOptions, do display.AllocationDisplayOptions, qo q
 
 	table := tview.NewTable()
 
-	var allocations map[string]kubecost.Allocation
+	var allocations map[string]opencost.Allocation
 	var allocMutex sync.Mutex
 	var lastUpdated time.Time
