+++ pkg/query/options.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/kubecost/kubectl-cost/pkg/cmd/utilities"
-	"github.com/opencost/opencost/pkg/log"
+	"github.com/opencost/opencost/core/pkg/log"
 
 	"k8s.io/client-go/rest"
